backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.3
shap>=0.44.0
joblib>=1.3.0
pydantic>=2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests>=2.31.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
optuna>=3.5.0
plotly>=5.17.0

# Authentication dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.13.0

# Enhanced security dependencies (optional - install with requirements_security.txt)
# redis>=5.0.0
# argon2-cffi>=23.1.0
# httpx>=0.25.0 